Many small fixes. Now we've moved to post leader election phase. 1. Temporary hack with int64_t. Will eliminate -1 situation in the future. 2. Fix some uninitialized variables that will cause wrong behavior of leader election. 3. Update some node states that previously forget to update. 4. Add many debugging output.
Now, the program can successfully elect a leader. However, the program
segfaults when the new leader want to recover some information.
Shouldn't take too long to fix it. Two more functions to go!!!

+++ src/include/leader-election/view_controller.h
@@ -24,7 +24,8 @@
 // proposer and learner can be treated as an in-memory structure.
 //
 
-typedef uint32_t pnum_t;
+/*typedef uint32_t pnum_t;*/
+typedef int64_t pnum_t;
 typedef enum leader_election_msg_type_t{
     LELE_PREPARE = 0,
     LELE_PREPARE_ACK = 1,
