+++ src/include/util/common-structure.h
@@ -20,7 +20,7 @@
 #include <stdint.h>
 #include <sys/time.h>
 
-typedef uint32_t node_id_t;
+typedef int64_t node_id_t;
 typedef uint32_t req_id_t;
 typedef uint32_t view_id_t;
